% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Initialization.R
\name{abscissae}
\alias{abscissae}
\title{Abscissae function}
\usage{
abscissae(h,domain,x1 = NULL,xk =NULL,x0=0,nmesh=5,min_step=0.001,relax_factor=5)
}
\arguments{
\item{h}{The log of the density function}

\item{domain}{A vector of length 2 giving the domain (left bound, right bound) of the density function}

\item{x1}{The leftmost node}

\item{xk}{The rightmost node}

\item{x0}{An initial point for searching for suitable x1 and xk. If provided by the user based on the knowledge of the density function, then it would be faster to find x1 and xk}

\item{nmesh}{The number of nodes in the initial grid}

\item{min_step}{The minimum step allowed in the adaptive-step search algorithm, if the variance is extremely small, then the used can decrease the min_step, yet it would be slower}

\item{relax_factor}{The relaxation factor in the adaptive-step search algorithm}
}
\description{
Looks for suitable x1 and xk if they are not given, this can ensure the adaptive algorithm is not biased as well as avoiding numerical issues. It can then generate the initial grid, the default number of nodes is 5. Depending on the domain, different algorithms are used, but they are all adaptive. The relaxation factor and minimum step can be defined as well.
}

