% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add together two numbers and add 6 to the sum.}
\usage{
add(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The sum of \code{a} and \code{b}.
}
\description{
Add together two numbers and add 6 to the sum.
}
\examples{
add(1, 1)
add(10, 1)
}

