\name{ARS-package}
\alias{ARS-package}
\alias{ARS}
\docType{package}
\title{
\packageTitle{ARS}
}
\description{
\packageDescription{ARS}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ARS}

Maintainer: \packageMaintainer{ARS}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ARS)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ARS}}
%\section{Documentation Index}{\packageIndices{ARS}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
