% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\docType{package}
\name{ars_package}
\alias{ars_package}
\alias{ars_package-package}
\title{ars: A package for completing stat243.}
\description{
The ars package provides one important function: ars.
}
\section{ars functions}{

ars
}

