% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{beta_u_x}
\alias{beta_u_x}
\title{Normalized Beta and Weights}
\usage{
beta_u_x(m, b, bounds)
}
\arguments{
\item{m}{numeric vector of slope}

\item{b}{numeric vector of intercept}

\item{bounds}{numeric vector of abscissae bounds for function}
}
\value{
A list of length 2. \code{beta} numeric vector of normalized beta,
\code{w} numeric vector of area weights.
}
\description{
Calculates the normalized beta and weights for sampling \eqn{u_k}.
}
