% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{boundary_check}
\alias{boundary_check}
\title{Boundary Check}
\usage{
boundary_check(x, bound)
}
\arguments{
\item{x}{numeric vector}

\item{bounds}{numeric vector of length 2, lower bound and upper bound}
}
\value{
\code{TRUE} if \code{x} is within given bounds, \code{FALSE} otherwise.
}
\description{
Checks whether values are within given bounds.
}
