% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{boundary_warning}
\alias{boundary_warning}
\title{Boundary Flag}
\usage{
boundary_warning(boolean_check)
}
\arguments{
\item{boolean_check}{boolean}
}
\value{
Gives a warning message if the \code{boolean_check} is \code{FALSE}.
}
\description{
Whether the boundary condition is being checked.
}
