% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_deriv}
\alias{calc_deriv}
\title{calc_deriv}
\usage{
calc_deriv(x, f, lower, upper, ...)
}
\arguments{
\item{x}{a specific x point to find the derivative at}

\item{f}{a function to calculate the derivative of}

\item{lower}{the lower bound domain to check with f}

\item{upper}{the upper bound domain to check with f}

\item{...}{additional variables needed in the f function}
}
\value{
a numeric value that represents the derivative at a certain point
}
\description{
Calculate the derivative of function f: f'(x) with an increment of machine epsilon
}
\examples{
calc_deriv(5, dnorm, -5, 5, mean=5, sd=1)
}
