% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hulls.R
\name{calculate_tangents}
\alias{calculate_tangents}
\title{Calculates the intersections of the tangents of a set of
abscissae for a given density}
\usage{
calculate_tangents(abscissae, density)
}
\arguments{
\item{abscissae}{vector of points at which to calculate tangents & intersections}
}
\value{
y values of the intersection points of the tangents
}
\description{
Calculates the intersections of the tangents of a set of
abscissae for a given density
}
