% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{cavitySearch}
\alias{cavitySearch}
\title{Search a function for non-log concavity}
\usage{
cavitySearch(f, f_params, x)
}
\arguments{
\item{f}{The function to evaluate, as an expression.}

\item{f_params}{A \code{list} of any associated parameters of \code{f}.}

\item{x}{The value at which to evaluate the derivative of \code{f}.}
}
\value{
logical that evaluates to TRUE if the density is log-concave at the
 specified value, and FALSE otherwise.
}
\description{
\code{cavitySearch} checks that a particular function is
 log-concave
}
\keyword{internal}
