% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_density_convergence.R
\name{check_density_convergence}
\alias{check_density_convergence}
\title{Check convergence of a density}
\usage{
check_density_convergence(density, lower, upper)
}
\arguments{
\item{density}{The unnormalized target density}

\item{lower/upper}{The lower / upper bound of the domain, where the points are sampled from}
}
\value{
A finite number the unnormalized density converge to.
}
\description{
Check if a given density has finite integral over the sampling range. If it diverges, error is reported.
}

