% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_interpolconcave}
\alias{check_interpolconcave}
\title{check_interpolconcave}
\usage{
check_interpolconcave(x, f)
}
\arguments{
\item{x}{x points}

\item{f}{represent f(x)}
}
\value{
a boolean TRUE or FALSE. A TRUE means the function is concave
}
\description{
given a (sorted) set of points,
check if the interpolating function is concave
}
\examples{
log_f <- function(x) log(dnorm(x))
x <- seq(-3, 3, by=1)
y <- log_f(x)
check_interpolconcave(x, y)
}
