% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logconcave.R
\name{check_local_concave}
\alias{check_local_concave}
\title{Check local concavity}
\usage{
check_local_concave(u, l, xvec, halfrange = 1e-06)
}
\arguments{
\item{u}{The piecewise linear upper hull functions (tangent lines)}

\item{l}{The piecewise linear lower squeezing functions (secant lines)}

\item{xvec}{The abscissae of the chosen x values, the left/right of which are to be insepected}

\item{halfrange}{The distance of the inspected points from the chosen abscissae}
}
\value{
FALSE if any tangent value is lower; TRUE if all tangent values are higher.
}
\description{
Check the local concavity of a function by taking left/right points of a vector of chosen abscissae, and compare the tangent/secant values. If any of the inspected points has larger secant values than tagent, then the function is not concave.
}

