% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{check_param}
\alias{check_param}
\title{Adaptive-Rejection Sampling Parameters Test}
\usage{
check_param(n, g, h, h_prime, D_left, D_right, k, step, center)
}
\arguments{
\item{n}{The number of desired samples}

\item{g}{The density function of interest}

\item{h}{The log of g (users can provide either g or h)}

\item{h_prime}{H prime function (optional)}

\item{D_left}{The desired left end of domain (optional), default = -Inf}

\item{D_right}{The desired right end of domain (optional), default = Inf}

\item{k}{The number of desired initial abscissae (optional), default = 10}

\item{step}{The desired width (optional), default = 3}

\item{center}{the desired center (optional), default = 0}
}
\value{
If an error is encounter, provides the user some feedback regarding the source of error.
}
\description{
The objective of this function is to test the parameter inputs of the user. It checks
for validity of the inputs and returns comments if an error is encountered with the inputs.
}
