% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_positive}
\alias{check_positive}
\title{check_positive}
\usage{
check_positive(f, var_min, var_max)
}
\arguments{
\item{f}{a function to search across}

\item{var_min}{the lower bound domain to check with f}

\item{var_max}{the upper bound domain to check with f}
}
\value{
a boolean value that represents whether a function is only positive
}
\description{
Check whether a function has negative values
}
\examples{
check_positive(dnorm, -Inf, Inf)
}
