% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_support_boundaries.R
\name{check_support_boundaries}
\alias{check_support_boundaries}
\title{Check the boundary is supportive to the function f}
\usage{
check_support_boundaries(f, lower, upper)
}
\arguments{
\item{f}{The density function (which could be unnormalized)}

\item{lower/upper}{The boundaries provided by the user}
}
\value{
If the unnormalized density is positive and finite within the bounds; otherwise, the program is aborted
}
\description{
Check the boundary is supportive to the function f
}

