% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkAccept.R
\name{checkAccept}
\alias{checkAccept}
\title{Helper function to check if the point is accepted as one of the generating
samples}
\usage{
checkAccept(x, z, g)
}
\arguments{
\item{x}{A vector of x values of all points}

\item{z}{A vector of z values of all points and we should be able to get
the index of z}

\item{g}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}
}
\value{
A vector of boolean value TRUE or FALSE indicating whether it's
accepted and the value of x*
}
\description{
Helper function to check if the point is accepted as one of the generating
samples
}

