% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkLC.R
\name{checkLC}
\alias{checkLC}
\title{Helper function to check the log concavity}
\usage{
checkLC(f)
}
\arguments{
\item{f}{A function that requires the log concavity check}
}
\value{
boolean value TRUE or FALSE to indicate the log concavity
}
\description{
Helper function to check the log concavity
}

