% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{checkThat}
\alias{checkThat}
\title{Check arguments of the call to ars.}
\usage{
checkThat(f, f_params, starting_values, sample_size)
}
\arguments{
\item{f}{A function representing the target sampling distribution.}

\item{f_params}{A \code{list} of any associated parameters of \code{f}.}

\item{starting_values}{The initial values for the k abscissae.  If NULL,
values will be determined using the bounds of the log-distribution where the
first derivative is positive.}

\item{sample_size}{The desired number of samples in the final output.}
}
\value{
A warning or error if certain conditions are detected
}
\description{
\code{checkThat} performs tests and checks to ensure
 functionality of the adaptive rejection sampling algorithm.  For example, if
 no starting values are supplied, \code{checkThat} will determine viable
 starting values for the initialization step.
}
\keyword{internal}
