% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choose2.R
\name{choose2}
\alias{choose2}
\title{Helper function to choose two starting points}
\usage{
choose2(g, D)
}
\arguments{
\item{g}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x)).}

\item{D}{A vector with two elements indicates the domain of the sample
generation.}
}
\value{
two starting points.
}
\description{
Helper function to choose two starting points
}

