% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{computeZ}
\alias{computeZ}
\title{computeZ}
\usage{
computeZ(abscissae_summary)
}
\arguments{
\item{abscissae_summary}{the abscissae matrix with the three columns being the x's,
h(x)'s, and derivatives of the h(x)'s respectively. There should be no Inf's
in the third column}
}
\value{
a numeric vector consists of the x-coordinates of the intersection points
(i.e. z in paper by Gilks et al.)
}
\description{
A function that computes the intersections of the tangent lines, provided the
abscissae matrix
}

