% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lowhull}
\alias{create_lowhull}
\title{create_lowhull}
\usage{
create_lowhull(x, f)
}
\arguments{
\item{x}{x points}

\item{f}{represent f(x)}
}
\value{
a function that calculates the lowerhull for a function
}
\description{
given a (sorted) set of points, create the lower hull function
}
\examples{
log_f <- function(x) log(dnorm(x))
x <- seq(-3, 3, by=1)
y <- log_f(x)
l <- create_lowhull(x, y)
l(0)
}
