% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_upphull}
\alias{create_upphull}
\title{create_upphull}
\usage{
create_upphull(x, f)
}
\arguments{
\item{x}{x points}

\item{f}{represent f(x)}
}
\value{
a function that calculates the upperhull for a function
}
\description{
given a (sorted) set of points, create the upper hull function
}
\examples{
log_f <- function(x) log(dnorm(x))
x <- seq(-3, 3, by=1)
y <- log_f(x)
u <- create_upphull(x, y)
u(0)
}
