% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_deriv.R
\name{der_fun}
\alias{der_fun}
\title{Calculates the derivatives at specified order}
\usage{
der_fun(f, x, ..., order = 1, delta = 10^-2, sig = 6)
}
\arguments{
\item{f}{The function to take derivative for}

\item{x}{The abscissa at which the derivative is taken}

\item{order}{The order of the derivative}

\item{delta}{The infinidesimal piece for calculation}
}
\value{
The derivative value at point x for function h of the specified order
}
\description{
Calculate the derivative locally for a function of specific order
}

