% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deriv}
\alias{deriv}
\title{Calculates the derivative of the log density, using the chain rule
where f(x) = log(x) and g(x) = pdf, so d/dx f(g(x)) = g_prime(x)/g(x)}
\usage{
deriv(density, x, epsilon = 0.001)
}
\arguments{
\item{density}{closure of the density of interest}

\item{x}{value of point at which to integrate}

\item{epsilon}{change value used to calculate derivative}
}
\value{
value of derivative of log-density at a given point
}
\description{
Calculates the derivative of the log density, using the chain rule
where f(x) = log(x) and g(x) = pdf, so d/dx f(g(x)) = g_prime(x)/g(x)
}
