% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw sampling point candidates}
\usage{
draw_sample(vertices, z, num_of_samples = 1)
}
\arguments{
\item{vertices}{The vertices matrix that describe the current vertices}

\item{z}{The vector of abscissa of the intersection points of tangent lines}

\item{num_of_samples}{The number of samples to draw}
}
\value{
The candidate samples that are drawn from the upper hull functions, which will soon be tested and accepted/rejected
}
\description{
Draw samples from the density described by the upper hull function
}

