% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{duplication_check}
\alias{duplication_check}
\title{Duplication Check}
\usage{
duplication_check(x, y, L, eps, dx)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{L}{length}

\item{eps}{epsilon}

\item{dx}{mesh size}
}
\value{
A boolean vector, \code{TRUE} if x and y do not contain duplicates,
\code{FALSE} otherwise.
}
\description{
Checks whether vectors contain duplicate values.
}
