% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Envelop.R
\name{envelop}
\alias{envelop}
\title{Envelope function}
\usage{
envelop(h,x,domain)
}
\arguments{
\item{h}{The log of the density function}

\item{x}{The nodes, found by abscissae()}

\item{domain}{A vector of length 2 giving the domain (left bound, right bound) of the density function}
}
\description{
Calculates the upper hull of h. The returned value is a matrix with each row storing (h'(x), h(x_i),x_i,xmin,xmax) for each node. In this way, each upper hull is uniquely defined, as well as its corresponding domain.
}

