% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_inverse_cdf}
\alias{eval_inverse_cdf}
\title{Numerically inverts a given CDF}
\usage{
eval_inverse_cdf(prob, cdf)
}
\arguments{
\item{prob}{probability value at which to invert the CDF}

\item{cdf}{closer of the cdf of interest}
}
\value{
value of the inverse of the cdf for a given probability
}
\description{
Numerically inverts a given CDF
}
