% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_l.R
\name{eval_l}
\alias{eval_l}
\title{eval_l}
\usage{
eval_l(x1, x, hfun)
}
\arguments{
\item{x1}{points that want to evalue lower bound}

\item{x}{abscissae}

\item{hfun}{log density function of interest, not need to be normalized.}
}
\value{
lower bound value
}
\description{
Evaluate the lower bound value
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

