% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_u.R
\name{eval_u}
\alias{eval_u}
\title{eval_u}
\usage{
eval_u(x.temp, u)
}
\arguments{
\item{x.temp}{points that want to evalue upper bound}

\item{u}{list contains coefficients and breaks}
}
\value{
upper bound value
}
\description{
Evaluate the upper bound value
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

