% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_deriv.R
\name{evaluate_deriv}
\alias{evaluate_deriv}
\title{Calculates first order derivative}
\usage{
evaluate_deriv(h, x, diff = 10^-12)
}
\arguments{
\item{h}{A differentiable function}

\item{x}{The abscissa at which the derivative is taken}

\item{diff}{The infinidesimal piece for calculation}
}
\value{
The derivative value at point x for function h
}
\description{
Numerically calculate the derivative locally for a continuous differentiable function
}

