% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_AcceptRejectx.R
\name{faux_AcceptRejectx}
\alias{faux_AcceptRejectx}
\title{Helper function to check if the point is accepted as one of the generating
samples}
\usage{
faux_AcceptRejectx(inp_xvec, inp_zvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of x values of all points}

\item{inp_zvec}{A vector of z values of all points and we should be able to get 
the index of z}

\item{inp_gfun}{A function user wants to generate samples from. This function is 
used to calculate \eqn{h(x)=\log(g(x))}{h(x) = ln(g(x))}}
}
\value{
A vector of boolean value TRUE or FALSE indicating whether it's 
    accepted and the value of \eqn{x^{*}}{x*}.
}
\description{
Helper function to check if the point is accepted as one of the generating
samples
}

