% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_hPrimex.R
\name{faux_hPrimex}
\alias{faux_hPrimex}
\title{Helper function to get first derivative of \eqn{h(x)}}
\usage{
faux_hPrimex(inp_gfun, inp_xvec)
}
\arguments{
\item{inp_gfun}{A function of \eqn{x} which the user wants to generate
samples from.
This function is used to calculate \eqn{h(x)=\log(g(x))}{h(x) = ln(g(x))}.}

\item{inp_xvec}{A number indicates the x-axis of the point}
}
\value{
A numeric vector of length equal to \code{inp_xvec}.
The elements of the returned value are equal to the first derivative of
\eqn{h(x)} at the points in \code{inp_xvec}.
}
\description{
Helper function to get first derivative of \eqn{h(x)}
}

