% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_hPrimex.R
\name{faux_hx}
\alias{faux_hx}
\title{Helper function to get first derivative of h(x)}
\usage{
faux_hx(g)
}
\arguments{
\item{g}{A function of x which the user wants to generate samples from. 
This function is used to calculate h(x)=ln(g(x)).}

\item{x}{A number indicates the x-axis of the points}
}
\value{
the first derivative of h(x) at the point x
}
\description{
Helper function to get first derivative of h(x)
}

