% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_InitChoose3.R
\name{faux_InitChoose3}
\alias{faux_InitChoose3}
\title{Helper function to choose two starting points}
\usage{
faux_InitChoose3(inp_gfun, D)
}
\arguments{
\item{inp_gfun}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}

\item{inp_Dvec}{A vector with two elements indicates the support domain of
the sample generation.}
}
\value{
two starting points.
}
\description{
Helper function to choose two starting points
}

