% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Lkx.R
\name{faux_Lkx}
\alias{faux_Lkx}
\title{Helper function to get the lower bound linear function \eqn{l_{k}(x)}{lk(x)}.}
\usage{
faux_Lkx(inp_xvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of x values of all points and we should be able to get the
index of \eqn{x}}

\item{inp_gfun}{A function user wants to generate samples from. This function is
use to calculate \eqn{h(x)=\log(g(x))}{h(x) = ln(g(x))}}
}
\value{
A list of functions.
The length of the list is one greater than the length of the input
\code{inp_xvec}.
Each element of the list is a piece of the piecewise function
\eqn{l_{k}(x)}{lk(x)}, which forms the lower hull of the function \eqn{h(x)}.
}
\description{
Helper function to get the lower bound linear function \eqn{l_{k}(x)}{lk(x)}.
}

