% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Skx.R
\name{faux_Skx}
\alias{faux_Skx}
\title{Helper function to get function \eqn{s_k(x)}}
\usage{
faux_Skx(z_int, uFun_out)
}
\arguments{
\item{z_int}{A vector of z values of all points and we should be able to get 
the index of z}

\item{uFun_out}{A list of functions, the output from uFun.}
}
\value{
the function $s_{k}(x)$ for all x in the domain of g.
}
\description{
Helper function to get function \eqn{s_k(x)}
}

