% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_init.R
\name{find_init}
\alias{find_init}
\title{Find the initial points}
\usage{
find_init(g, my_total_range)
}
\arguments{
\item{g}{the density function used to perform the rejection sampling, where g(x) = cf(x). f(x) is the density that we actually want to sample points from}

\item{my_total_range}{support of the probability density distribution}
}
\description{
This function finds the initial points of the adptive rejection sampling.
}

