% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_intercept}
\alias{find_intercept}
\title{Find the x values of the intercept.}
\usage{
find_intercept(x_pts)
}
\arguments{
\item{x_pts}{points selected initially to find the tangent line, abscissae,
T_k.}
}
\value{
the x values of the intercept.
}
\description{
This function returns the x values of the intercept.
}
\details{

}
\examples{
a = c(-3, -0.2, 2)
find_intercept(a)

}
