% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_l}
\alias{find_l}
\title{Calculate the lower bound.}
\usage{
find_l(x, x_pts)
}
\arguments{
\item{x}{points randomly selected to plot the bounds.}

\item{x_pts}{points selected initially to find the tangent line, abscissae,
T_k.}
}
\description{
This function calculates the lower bound.
}
\details{
In addition to x and x_pts, \code{find_l} uses an additional parameter
during its caculations: l. l is the corresponding y values for x values
at x_pts.
}
\examples{
a = c(-3, -0.2, 2)
x = seq(-4, 3.9, length.out = 500)
find_l(x, a)

}
