% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find mode of a density function}
\usage{
find_mode(density, lb = -Inf, ub = Inf)
}
\arguments{
\item{density}{The density function to sample from}

\item{lb/ub}{The range of the density function within which the mode is being seeked. The default values are set to plus/minus infinity.}
}
\value{
A numeric vector of information found in the procedure. The 1st element is the abscissa of the mode in the range. The 2nd element is the density at the mode. The 3rd/4th elements are the updated(shrinked) lower/upper bound of the sampling range, if the input range is infinite.
}
\description{
Calculate the abscissa of the mode of the input density function within the given range. It utilizes Brent when either one of the lower/upper bounds is finite, and BFGS when both of them infinite. The function also shrinks sampling range by assigning a finite value to the infinite bound(s) if there exist one.
}

