% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_s}
\alias{find_s}
\title{Find S.}
\usage{
find_s(x, x_pts)
}
\arguments{
\item{x}{points randomly selected to plot the bounds.}

\item{x_pts}{points selected initially to find the tangent line, abscissae,
T_k.}
}
\description{
This function finds s.
}
\examples{
a = c(-3, -0.2, 2)
find_s(c(-1, 0, 1), a)

}
