% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_u}
\alias{find_u}
\title{Calculate the upper bound.}
\usage{
find_u(x, x_pts)
}
\arguments{
\item{x}{points randomly selected to plot the bounds.}

\item{x_pts}{points selected initially to find the tangent line, abscissae,
T_k.}
}
\description{
This function calculates the upper bound.
}
\details{
In addition to x and x_pts, \code{find_u} uses two additional parameters
during its caculations: u and norm_const. u is the corresponding y values for
x values at x_pts. Norm_const is the area under the curve.

#' a = c(-3, -0.2, 2)
x = seq(-4, 3.9, length.out = 500)
find_u(x, a)
}
