% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{finite_check}
\alias{finite_check}
\title{Finite Check}
\usage{
finite_check(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
\code{TRUE} if \code{x} and \code{y} are finite, \code{FALSE} otherwise.
}
\description{
Checks for NaNs and infinities.
}
