% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_xt.R
\name{gen_xt}
\alias{gen_xt}
\title{Piecewise linear upper hull}
\usage{
gen_xt(cs_normalized_integral_values, integral_values, my_points, my_values,
  my_slopes, domains)
}
\arguments{
\item{cs_normalized_integral_values}{normalized integral exponential upper hull}

\item{integral_values}{piecewise integrated upper hull function}

\item{my_points}{sorted points in T, which contains k abscissae in the domain D}

\item{my_values}{evaluated h at x_intercepts, where h = log g(x). g is the density function used to perform the rejection sampling}

\item{my_slopes}{slope of h evaluated at x_intercepts; they are the outputs of dh function calculated at x1 through xk}

\item{domains}{domain D, which contains the lower bound and the upper bound}
}
\description{
This function returns the piecewise linear upper hulls from the tangents to h(x) at the abscissae x_intercepts in Tk
}

