% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_u.R
\name{generate_u}
\alias{generate_u}
\title{genrate_u}
\usage{
generate_u(x, hfun.x, hfun_deriv.x, start, end)
}
\arguments{
\item{x}{initial points}

\item{hfun.x}{log function of initial points}

\item{hfun_deriv.x}{derivative of log function of initial points}

\item{start}{lower bound}

\item{end}{upper bound}
}
\value{
list contains coefficients and breaks
}
\description{
Generate coefficients and breaks
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

