% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genu.R
\name{genu}
\alias{genu}
\title{genu}
\usage{
genu(mat, start, end)
}
\arguments{
\item{mat}{A k by 3 matrix. First column is initial value. Second column is the log function value of initial value. Thrid column is the derivative of the log function of initial value.}

\item{start}{lower bound}

\item{end}{upper bound}
}
\value{
list contains coefficients and breaks
}
\description{
Generate coefficients and breaks
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

