% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cdf}
\alias{get_cdf}
\title{Numerically calculates the CDF for a given density}
\usage{
get_cdf(density)
}
\arguments{
\item{density}{closure of the density of interest}
}
\value{
closure of the cdf which can be evaluated at a point
}
\description{
Numerically calculates the CDF for a given density
}
