% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecion.R
\name{get_intersection}
\alias{get_intersection}
\title{Get intersection points}
\usage{
get_intersection(vertices, lb, ub)
}
\arguments{
\item{vertices}{The vertices matrix that describe the current vertices}

\item{lb/ub}{The lower/upper bound of the sampling range}
}
\value{
A numeric vector that stores the absissae of intersection points and lb/ub, in ascending order
}
\description{
Calculate and combines intersection points of all tangent lines in the current setup, and append the lower/upper bound of the sampling range
}

