% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hPrime.R
\name{hPrime}
\alias{hPrime}
\title{Helper function to get first derivative of h(x)}
\usage{
hPrime(inp_g, x)
}
\arguments{
\item{inp_g}{A function of x which the user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x)).}

\item{x}{A number indicates the x-axis of the points}
}
\value{
the first derivative of h(x) at the point x
}
\description{
Helper function to get first derivative of h(x)
}

