% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_abscissa}
\alias{init_abscissa}
\title{Initialize Abscissas}
\usage{
init_abscissa(h, a, b, init_k = 20)
}
\arguments{
\item{h}{log density function}

\item{a}{lower bound of function domain D (could be -Inf)}

\item{b}{upper bound of function domain D (could be Inf)}

\item{init_k}{number of xs in the initial abscissa, default is 20}
}
\value{
a vector consisting of all xs in the abscissa
}
\description{
Initialize the abscissas (x vector), based on the function domain D.
If D is bounded, select xs with equal distances, excluding the bounds
If D is left/right unbounded, select first/last x based on h's derivative values
}
