% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_vertices}
\alias{init_vertices}
\title{Initialize the vertices matrix}
\usage{
init_vertices(h, lb, ub, condition, mode)
}
\arguments{
\item{h}{The log-density function}

\item{condition}{The sampling condition which was decided at the initial setup. It takes on values: 1 - uniform distribution, 2 - left truncated, 3 - right truncated, and 4 - a regular density}

\item{mode}{The abscissa of the mode of the density}

\item{lb/ub}{The lower/upper bound of the sampling range}
}
\value{
The initial vertices matrix that stores the information of the points where we define the tangent upper hull and secant lower squeezing functions
}
\description{
Initialize the vertices matrix as a starting point for further sampling
}

