% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{initialization_step}
\alias{initialization_step}
\title{Initialize Step}
\usage{
initialization_step(h, a, b)
}
\arguments{
\item{h}{log density function}

\item{a}{lower bound of function domain D (could be -Inf)}

\item{b}{upper bound of function domain D (could be Inf)}
}
\value{
a list, the main data structure we will work with, containing
x, h(x), d'(x), z, and other information
}
\description{
Initialize the data structure we maintain given the log density function h as input,
as described in 2.2.1.
We calculate the initial abscissa, h(x) vector, h'(x) vector,
and z vector needed for further computation
}
