% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{initialize}
\alias{initialize}
\title{initialize}
\usage{
initialize(funx, D = c(-Inf, Inf), ...)
}
\arguments{
\item{D}{a vector of 2 points to represent the lower and upper bounds}

\item{funcx}{represent f(x) we want to sample from}

\item{y}{may represent different y values from f(x)}
}
\value{
a R6 class that contains the relevant information we need
}
\description{
create the ARS object by calling $new()
}
