% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitTk.R
\name{initTk}
\alias{initTk}
\title{InitTk}
\usage{
initTk(hfun, hfun_deriv, start, end)
}
\arguments{
\item{hfun}{log density function of interest, not need to be normalized.}

\item{hfun_deriv}{the derivative of log density function.}

\item{start}{lower bound.}

\item{end}{upper bound.}
}
\value{
A k by 3 matrix. First column is initial value. Second column is the log function value of initial value. Thrid column is the derivative of the log function of initial value.
}
\description{
Generate the initial value for Adaptive rejection sampling
}
\examples{

#Initial value for standard normal distribution
set.seed(0)
hfun <- function(x) dnorm(x, log=TRUE)
h <- 1e-8
hfun_deriv <- function(x) (hfun(x + h) - hfun(x - h))/(2*h)
InitTk(hfun, hfun_deriv, -Inf, Inf)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

