% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integ_expinterpol}
\alias{integ_expinterpol}
\title{integ_expinterpol}
\usage{
integ_expinterpol(x, f, y = x)
}
\arguments{
\item{x}{x points}

\item{f}{represent f(x)}

\item{y}{may represent different y values from f(x)}
}
\value{
a function that calculates the lowerhull for a function
}
\description{
integrate the exponential of interpolating function for each interval given a (sorted) set of points
}
