% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integration_check}
\alias{integration_check}
\title{Integration Check}
\usage{
integration_check(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
Throws an error if \code{x} + \code{y} is less than or equal to 0.
}
\description{
Checks whether the integral of s(x) is positive
}
