% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{intercept_z_j}
\alias{intercept_z_j}
\title{Z_j Intercept}
\usage{
intercept_z_j(x, h, dh, bounds)
}
\arguments{
\item{x}{numeric vector of x values to evaluate the tangents at}

\item{h}{concave function h(x) evaluated at x}

\item{dh}{derivative of h(x) evaluated at x}
}
\value{
Numeric vector of where the tangents to \eqn{u_k} intersect,
including bounds values.
}
\description{
Calculates the intercepts of \eqn{u_k(x)}, a piecewise linear
upper hull formed by the tangents to h(x).
}
