% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logconcave.R
\name{is_logconcave_shape}
\alias{is_logconcave_shape}
\title{Check the sampling condition based on shape}
\usage{
is_logconcave_shape(h, x_lo, x_hi, mode, ...)
}
\arguments{
\item{h}{The log-density}

\item{mode}{The abscissa of the mode of the density in the given range}

\item{x_lo/x_hi}{The lower and upper bounds of the sampling range}
}
\value{
The sampling condition coded in number 1-4. 1 - uniform distribution, 2 - left truncated, 3 - right truncated, and 4 - a regular density
}
\description{
Decide the sampling condition of the procedure based on the shape and implied log-concavity
}

