% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lk}
\alias{lk}
\title{Lower envelope function for h (l_k in formula (4))}
\usage{
lk(x_value, data)
}
\arguments{
\item{x_value}{the position to evaluate the u_k function}

\item{data}{the maintained data structure for the overall process}
}
\value{
the value of the lower envelope for l_k(x_value)
}
\description{
Calculating the lower envelope function at point x_value,
given the overall maintained data structure
}
