% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{log_concavity_check}
\alias{log_concavity_check}
\title{Log-concave Check}
\usage{
log_concavity_check(bound, L, x)
}
\arguments{
\item{bound}{numeric}

\item{L}{numeric}

\item{x}{numeric vector}
}
\value{
Gives an error message if the function f is not log-concave.
}
\description{
Checks whether the function f is log-concave.
}
