% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_piecewise.R
\name{lower_piecewise}
\alias{lower_piecewise}
\title{Piecewise lower hull}
\usage{
lower_piecewise(x, x_values, y_values, domains)
}
\arguments{
\item{x}{x-coordinate value}

\item{x_values}{sorted points in T, which contains k abscissae in the domain D}

\item{y_values}{evaluated h at x_values, where h = log g(x). g is the density function used to perform the rejection sampling}

\item{domains}{domain D, which contains the lower bound and the upper bound}
}
\description{
This function returns the piecewise linear lower hull formed from the chords between adjacent abscissae in Tk
}
\examples{
lower_piecewise(x, x_values, y_values, domains)
lower_piecewise
}

