% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{lowerHull}
\alias{lowerHull}
\title{Compute the Lower Hull}
\usage{
lowerHull(x, x_abs, f, f_params = NULL, supp = c(-Inf, Inf), hx = NULL)
}
\arguments{
\item{x}{A \code{numeric} vector representing x values.}

\item{x_abs}{A \code{numeric} vector of length \code{k} for \code{k > 1},
representing the abscissae.}

\item{f}{A function representing the target sampling distribution.}

\item{f_params}{A \code{list} of any associated parameters of \code{f}.}

\item{supp}{The support of \code{f}, as a two-membered \code{numeric} vector.
Default is \code{c(-Inf, Inf)}}

\item{hx}{The log of the density \code{f} at point x}
}
\value{
A named \code{list} containing the y-values of the lower hull.
}
\description{
\code{lowerHull} calculates the y-values corresponding to the
lower hull formed from the tangents to a function.
}
\keyword{internal}
