% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{max_iter_check}
\alias{max_iter_check}
\title{Maximum Iteration Check}
\usage{
max_iter_check(i, max_iter)
}
\arguments{
\item{i}{current iteration}

\item{max_iter}{maximum number of iterations}
}
\value{
Throws an error if the algorithm has reached the maximum number of
iterations.
}
\description{
Checks whether the algorithm has reached a maximum number of
iterations.
}
