% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{plot_sampdist}
\alias{plot_sampdist}
\title{plot_sampdist}
\usage{
plot_sampdist()
}
\value{
a ggplot object that represents the normalized envelope function.
}
\description{
plot the normalized envelope function
}
\examples{
n_samples = ars$new(dnorm)$sample(n=5000)$plot_sampdist()
}
