% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{plot_samples}
\alias{plot_samples}
\title{plot_samples}
\usage{
plot_samples()
}
\value{
a ggplot object that is a histogram and a blue line for the median sampled value
}
\description{
plot samples directly from ars method into a histogram with a blue line representing the median
}
\examples{
n_samples = ars$new(dnorm)$sample(n=5000)$plot_samples()
}
