% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{plot_u_l}
\alias{plot_u_l}
\title{plot_samples}
\usage{
plot_u_l()
}
\value{
a ggplot object that represents the upper and lower hulls
}
\description{
plot 100 points in our data with the respective upper and lower hull functions
}
\examples{
n_samples = ars$new(dnorm)$sample(n=5000)$plot_u_l()
}
