% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sampling.R
\name{rejection_test}
\alias{rejection_test}
\title{Rejection test}
\usage{
rejection_test(x_sampled,u,h)
}
\arguments{
\item{x_sampled}{a newly sampled x, from sample_one_point()}

\item{u}{the upper hull calculated using envelop(h,x,domain)}

\item{h}{the log density function}
}
\description{
Performs the rejection test for a newly sampled x if it fails the squeezing test.
}

