% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{sample}
\alias{sample}
\title{sample}
\usage{
sample(n = 1000, init_l = -1, init_r = 1, squeeze = TRUE)
}
\arguments{
\item{n}{represent the number of samples we want to create}

\item{init_l}{the left initial starting point for the ARS algorithm}

\item{init_r}{the right initial starting point for the ARS algorithm}

\item{squeeze}{a boolean implementing the squeezing test check or not.}
}
\value{
a vector of length n
}
\description{
samples n points
}
\examples{
n_samples = ars$new(dnorm)$sample(n=5000)
}
