% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_from_hull}
\alias{sample_from_hull}
\title{Generate a sample from a given density using inverse CDF method}
\usage{
sample_from_hull(hull)
}
\arguments{
\item{hull}{density of interest (a bounding hull in this context)}
}
\value{
sample from the density
}
\description{
Generate a sample from a given density using inverse CDF method
}
