% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_x.R
\name{sample_x}
\alias{sample_x}
\title{sample_x}
\usage{
sample_x(mu, y)
}
\arguments{
\item{mu}{list contains coefficients and breaks}

\item{y}{uniform random number}
}
\value{
x value
}
\description{
Sample the x value from S(x)
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

