% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{sampleEnv}
\alias{sampleEnv}
\title{Sample from the Rejection Envelope}
\usage{
sampleEnv(n, x_abs, f, f_params = NULL, supp = c(-Inf, Inf))
}
\arguments{
\item{n}{The number of samples}

\item{x_abs}{A \code{numeric} vector of length \code{k} for \code{k > 1},
representing the abscissae.}

\item{f}{A function representing the target sampling distribution.}

\item{f_params}{A \code{list} of any associated parameters of \code{f}.}

\item{supp}{The support of \code{f}, as a two-membered \code{numeric} vector.
Default is \code{c(-Inf, Inf)}}
}
\value{
A \code{numeric} vector containing the sampled values.
}
\description{
\code{sampleEnv} samples from the rejection envelope formed
by the piecewise upper hull for the given function
}
\keyword{internal}
