% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UpperCDFandInverse.R
\name{sampleUpper}
\alias{sampleUpper}
\title{sampleUpper}
\usage{
sampleUpper(n, CDFInverse)
}
\arguments{
\item{n}{a positive integer as the number of samples}

\item{CDFInverse}{an inverse CDF function of the upper hull}
}
\value{
a numeric vector sampled from upper hull, which has size n.
}
\description{
Sample from upper hull based on inverse CDF
}

