% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{sampling_step}
\alias{sampling_step}
\title{Rejection Sampling Step}
\usage{
sampling_step(sample_vector, data, h)
}
\arguments{
\item{sample_vector}{vector of x-values. Output of exp_sampling}

\item{data}{data structure output from the initialization_step function}

\item{h}{log density function}
}
\value{
output of sample data in our data structure
}
\description{
Given a sample output from the exp_sampling function, perform rejection
squeezing and rejection sampling steps. Output will be a structured list
of data to update
}
