% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sampling_x}
\alias{sampling_x}
\title{Sampling Function}
\usage{
sampling_x(N, beta, m, weights, bounds)
}
\arguments{
\item{N}{Number of samples}

\item{beta}{normalized beta given by \code{beta_u_x} function}

\item{m}{slope given by \code{slope_intercept_z_j} function}

\item{weights}{weights given by \code{beta_u_x} function}

\item{bounds}{intercepts given by \code{intercept_z_j} function}
}
\value{
A list of two vectors of length N of sampled x and J.
}
\description{
Samples from \eqn{s_k(x)} function distribution, select segment
of sampling based on the areas and use inverse CDF for sampling
}
