% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{search}
\alias{search}
\title{Search function}
\usage{
search(h,domain,x_start,max_x,min_step,relax_factor)
}
\arguments{
\item{h}{The log of the density function}

\item{domain}{A vector of length 2 giving the domain (left bound, right bound) of the density function}

\item{x_start}{The starting x value for the search}

\item{max_x}{The maximum search limit for x, if the mode is shifted beyond this limit, the user should define this parameter}

\item{min_step}{The minimum step allowed in the adaptive-step search algorithm, if the variance is extremely small, then the used can decrease the min_step, yet it would be slower}

\item{relax_factor}{The relaxation factor in the adaptive-step search algorithm}
}
\description{
Searches for a x0 such that h(x0) is finite and the gradient of h at x0 is attained. This is particular useful when the domain is unbounded, the mode is far away from 0 (yet unknown) and the variance is very small, i.e. the density function is very narrow. In such cases, it's even difficult to make a guess for an initial value to find x1 and xk.
}

