% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk.R
\name{sk}
\alias{sk}
\title{Piecewise integrals}
\usage{
sk(xj, hj, dhj, my_domains)
}
\arguments{
\item{xj}{sorted points in T}

\item{hj}{log density evaluated at point xj}

\item{dhj}{slopes of piecewise upper hull}

\item{my_domains}{intersection points of upper hull}
}
\description{
This function calculate the piecewise integral of the upper hull function.
}

