% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{slope_intercept_l_j}
\alias{slope_intercept_l_j}
\title{l_j Slope And Intercept}
\usage{
slope_intercept_l_j(x, h)
}
\arguments{
\item{x}{numeric vector of x values to evaluate at}

\item{h}{h(x) evaluated at x}
}
\value{
A list of length 2. \code{m} numeric vector of slopes, \code{b}
numeric vector of intercepts.
}
\description{
Calculates the slopes and intercepts of piecewise function
\eqn{l_j} at given x values. \eqn{l_j} is a piecewise linear lower hull
formed by adjacent chords.
}
