% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{squeezing_test}
\alias{squeezing_test}
\title{Squeezing test}
\usage{
squeezing_test(x_sampled,l,u)
}
\arguments{
\item{x_sampled}{a newly sampled x, from sample_one_point()}

\item{l}{the lower hull calculated using squeezing(h,x)}

\item{u}{the upper hull calculated using envelop(h,x,domain)}
}
\description{
Performs the squeezing test for a newly sampled x
}

