% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSample.R
\name{sSample}
\alias{sSample}
\title{Helper function to sample a value x*from $s_{k}(x)$}
\usage{
sSample(x, z, g)
}
\arguments{
\item{x}{A vector of x values of all points}

\item{z}{A vector of z values of all points and we should be able to get
the index of z}

\item{g}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}
}
\value{
a x value sampling from function $s_{k}(x)$
}
\description{
Helper function to sample a value x*from $s_{k}(x)$
}

