% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{subtract}
\alias{subtract}
\title{Add together two numbers and add 6 to the sum.}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers and add 6 to the sum.
}
\examples{
add(1, 1)
add(10, 1)
}

