% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{tanIntersect}
\alias{tanIntersect}
\title{Compute Tangent Lines and Intersections}
\usage{
tanIntersect(x, f)
}
\arguments{
\item{x}{A \code{numeric} vector of length \code{k} for \code{k > 1}.}

\item{f}{A function representing the target sampling distribution.}
}
\value{
A \code{numeric} vector of \code{k - 1} intersection points.
}
\description{
\code{tanIntersect} calculates the intersections of subsequent
tangent lines of a function.
}
\keyword{internal}
