% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uFun.R
\name{uFun}
\alias{uFun}
\title{Helper function to get the upper bound linear function $u_{k}(x)$}
\usage{
uFun(x, z, g)
}
\arguments{
\item{x}{A vector of x values of all points}

\item{z}{A vector of z values of all points and we should be able to get
the index of z}

\item{g}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}
}
\value{
the upper bound linear function $u_{k}(x)$ for x in the interval
[$z_{j-1}$,$z_{j}$]
}
\description{
Helper function to get the upper bound linear function $u_{k}(x)$
}

