% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uk}
\alias{uk}
\title{Upper envelope function for h (u_k in formula (2))}
\usage{
uk(x_value, data)
}
\arguments{
\item{x_value}{the position to evaluate the u_k function}

\item{data}{the maintained data structure for the overall process}
}
\value{
the value of the upper envelope for u_k(x_value)
}
\description{
Calculating the upper envelope function at point x_value,
given the overall maintained data structure
}
