% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Update.R
\name{update_grid}
\alias{update_grid}
\title{Updating step}
\usage{
envelop_density(h,x,domain)
}
\arguments{
\item{x}{The abscissae in the previous step}

\item{x_add}{An x value to be added to the abscissae}
}
\description{
UPdates the grid if both h and the gradient of h is evaluated at a sampled point.#'
}

