% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_method.R
\name{update_u}
\alias{update_u}
\title{Create Upper Hull function}
\usage{
update_u(vertices, lb, ub)
}
\arguments{
\item{vertices}{The vertices matrix that describe the current vertices}

\item{lb/ub}{The lower/upper boundaries of the sampling range}
}
\value{
A vectorized function that describes the updated upper hull function
}
\description{
Create(update) the upper hull function based on the current vertices matrix, utilizing the tangent line slopes
}

