% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_method.R
\name{update_vertices}
\alias{update_vertices}
\title{Update vertices}
\usage{
update_vertices(vertices, new_vertex, h)
}
\arguments{
\item{vertices}{The matrix structure to update}

\item{new_vertex}{The new point to add to the matrix}

\item{h}{The log-density}
}
\value{
A list, whose 1st element is the new vertices matrix, and the 2nd element is a boolean indicator that is TRUE when the boundaries should be shrinked
}
\description{
Add the new vertex each time the log-density is evaluated and a new point should be added to the vertices matrix
}

