% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{UpdateAccept}
\alias{UpdateAccept}
\title{Accept or reject sample
Note: added H, H_prime, Z as parameters to function#'}
\usage{
UpdateAccept(x_star, X, Z, H, H_prime, H_norm, P_cum, x_accept,
  length_accept, eval_h, eval_h_prime)
}
\description{
Accept or reject sample
Note: added H, H_prime, Z as parameters to function#'
}
