% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UpperCDFandInverse.R
\name{upperCDF}
\alias{upperCDF}
\title{upperCDF}
\usage{
upperCDF(x, abscissae_summary, xlow = -Inf, xhigh = Inf)
}
\arguments{
\item{x}{a numeric vector that takes values in [0,1] as the input for the CDF}

\item{abscissae_summary}{a numeric matrix comprised of abscissae together with h(x) and h'(x)}

\item{xlow}{a numeric value as the left bound of the support of g}

\item{xhigh}{a numeric value as the right bound of the support of g}
}
\value{
A numeric vector consists of the values of CDF for the upper hull at each abscissa.
}
\description{
Given a numeric vector and the abscissae matrix, function returns the CDF function of the upper hull.
}

