% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zFun.R
\name{zFun}
\alias{zFun}
\title{Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$}
\usage{
zFun(x, g)
}
\arguments{
\item{x}{A vector of x values of your points. The first element in the vector
has smaller x-coordinate.}

\item{g}{A fucntion user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}
}
\value{
intersection of tangents of $x_{j}$ and $x_{j+1}$ named $z_{j}$
}
\description{
Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$
}

