\name{ars-package}
\alias{ars-package}
\alias{ars}
\docType{package}
\title{
\packageTitle{ars}
}
\description{
\packageDescription{ars}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ars}
\packageIndices{ars}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ars}

Maintainer: \packageMaintainer{ars}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
