% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampler}
\usage{
ars(density, n, lb = -Inf, ub = Inf, ...)
}
\arguments{
\item{density}{A log-concave density function (could be unnormlalized)}

\item{n}{Number of samples}

\item{lb, ub}{the lower/upper bound of the domain where the points are sampled}
}
\value{
A vector of x-values that are sampled from the density input within the lb and ub
}
\description{
Generate n samples from a log-concave distribution function using adaptive rejection sampling
}

