% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampler}
\usage{
ars(density, n, lb = -Inf, ub = Inf, ...)
}
\arguments{
\item{density}{A log-concave density function (could be unnormlalized)}

\item{n}{Number of samples}

\item{lb, ub}{The lower / upper bound of the domain, where the points are sampled from}
}
\value{
A vector of values that are sampled from the density input within the lb and ub
}
\description{
Draw n samples from a log-concave density using adaptive rejection sampling
}

