% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Entrypoint for 'ars' package. Runs an adaptative
rejection sampling algorithm on a given probability density
and returns the specified number of samples.}
\usage{
ars(density, n_samples, k = 5)
}
\arguments{
\item{density}{A function closure of the density to be sampled from}

\item{n_samples}{The number of samples to sample from the distribution}

\item{k}{Number of initial points from which to construct tangents}
}
\value{
samples numeric vector of samples from the specified distribution
}
\description{
Entrypoint for 'ars' package. Runs an adaptative
rejection sampling algorithm on a given probability density
and returns the specified number of samples.
}
