% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Entrypoint for 'ars' package. Runs an adaptative
rejection sampling algorithm on a given probability density
and returns the specified number of samples.}
\usage{
ars(density, n_samples = 10, location = 0, scale = 1)
}
\arguments{
\item{density}{A function closure of the density to be sampled from.
Input functions must have similar behavior to dnorm, dexp, etc.}

\item{n_samples}{The number of samples to sample from the distribution}

\item{location}{starting point at which to initialize sampling points,
default 0. Should be chosen so that \code{location} is near the mode
of the distribution.}

\item{scale}{value at which to space initial points (default 1).
Should be chosen such that mode - (2\emph{scale), mode+(2}scale) is within
the support of the probability density.}
}
\value{
samples numeric vector of samples from the specified distribution
}
\description{
Entrypoint for 'ars' package. Runs an adaptative
rejection sampling algorithm on a given probability density
and returns the specified number of samples.
}
