% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\docType{class}
\name{ars}
\alias{ars}
\title{ARS class providing object with methods for communication with R6}
\format{\code{\link{R6Class}} object.}
\usage{
example <- ars$new(funx=dnorm, D=c(-Inf, Inf))
}
\value{
Object of \code{\link{R6Class}} with methods to generate adaptive rejection sampling
}
\description{
ARS class providing object with methods for communication with R6

ARS class providing object with methods for communication with R6
}
\section{Public Methods}{

\describe{
  \item{\code{sample(n, init_l, init_r, squeeze)}}{This method samples n points from using adaptive rejection sampling. User may pick two initalizing points, init_l and init_r which represent data points to start at. Useful if -1 and 1 are far from the pdf, i.e rnorm(mean=3000, sd=1)}
  \item{\code{f(x)}}{This method calculates our funx observation at a given x }
  \item{\code{plot_sampdist()}}{This method plots our normalized envelope function }
  \item{\code{calc_sampdist()}}{This method calculates the sampling distribution from a uniform distribution}
  \item{\code{plot_samples()}}{This method plots our samples in a histogramm, with a blue line highlighting the median value}
  \item{\code{s()}}{This method calculates the envelope or sampling function }}
}

\section{Private Methods}{

\describe{
  \item{\code{h(x)}}{This method calculates the log of our funx at a given point x}
  \item{\code{init_hprim()}}{This method intializes our hprim values}
  \item{\code{init_hval()}}{This method initalizes our h values}
  \item{\code{init_scdf()}}{This method initalizes our cdf under the upper hull}
  \item{\code{init_u()}}{This method initializes our upper hull values}
  \item{\code{init_y()}}{This method initalizes our y values }
  \item{\code{l(p)}}{This method is a function that returns the slope of a lowerhull line given a point p}
  \item{\code{u(p)}}{This method is a function that returns the slope of a upperhull line given a point p}
  \item{\code{sampl_exph()}}{This method samples from our exp(h values)}
  \item{\code{update(y, hy, hy_prim)}}{This method updates our private variables after an interation}}
}

\examples{
ex <- ars$new(dnorm, c(-Inf,Inf),mean=3, sd=1)
samples <- ex$sample(n=1000)
ex$plot_samples()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{self$f}}{Desnity function we want to sample from}

\item{\code{self$var_max}}{Upper bound on function we want to sample from}

\item{\code{self$var_min}}{Lower bound on function we want to sample from}

\item{\code{private$h_prims}}{A R6 namedVector that contains h'(x)}

\item{\code{private$h_vals}}{A R6 namedVector that contains h(x)}

\item{\code{private$s_cdfs}}{A R6 namedVector that contains each interval's CDF}

\item{\code{private$u_vals}}{A R6 namedVector that contains u(x)}

\item{\code{private$x}}{a vector of x values that are our sample values. Empty unless sample() run}

\item{\code{private$y}}{a vector of y values that begins to increase once sample() is run}

\item{\code{private$z}}{a vector of z values, or intersection points, that begins to increase once sample() is run}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{self$f}}{Desnity function we want to sample from}

\item{\code{self$var_max}}{Upper bound on function we want to sample from}

\item{\code{self$var_min}}{Lower bound on function we want to sample from}

\item{\code{private$h_prims}}{A R6 namedVector that contains h'(x)}

\item{\code{private$h_vals}}{A R6 namedVector that contains h(x)}

\item{\code{private$s_cdfs}}{A R6 namedVector that contains each interval's CDF}

\item{\code{private$u_vals}}{A R6 namedVector that contains u(x)}

\item{\code{private$x}}{a vector of x values that are our sample values. Empty unless sample() run}

\item{\code{private$y}}{a vector of y values that begins to increase once sample() is run}

\item{\code{private$z}}{a vector of z values, or intersection points, that begins to increase once sample() is run}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ars$new()}}
\item \href{#method-calc_sampdist}{\code{ars$calc_sampdist()}}
\item \href{#method-sample}{\code{ars$sample()}}
\item \href{#method-plot_samples}{\code{ars$plot_samples()}}
\item \href{#method-plot_sampdist}{\code{ars$plot_sampdist()}}
\item \href{#method-plot_u_l}{\code{ars$plot_u_l()}}
\item \href{#method-clone}{\code{ars$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$new(funx, D = c(-Inf, Inf), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{funx}}{a function we want to sample from}

\item{\code{D}}{a vector of length 2 of numbers for bounds of function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_sampdist"></a>}}
\subsection{Method \code{calc_sampdist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$calc_sampdist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$sample(n = 1000, init_l = -1, init_r = 1, squeeze = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_samples"></a>}}
\subsection{Method \code{plot_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$plot_samples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_sampdist"></a>}}
\subsection{Method \code{plot_sampdist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$plot_sampdist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_u_l"></a>}}
\subsection{Method \code{plot_u_l()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$plot_u_l()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
