% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\docType{class}
\name{ARS}
\alias{ARS}
\title{Class providing object with methods for communication with R6}
\format{\code{\link{R6Class}} object.}
\usage{
ARS
}
\arguments{
\item{funx}{a function we want to sample from}

\item{D}{a vector of length 2 of numbers for bounds of function}
}
\value{
Object of \code{\link{R6Class}} with methods to generate adaptive rejection sampling
}
\description{
Class providing object with methods for communication with R6
}
\section{Fields}{

\describe{
\item{\code{self$f}}{Function we want to sample from}

\item{\code{self$var_max}}{Upper bound on function we want to sample from}

\item{\code{self$var_min}}{Lower bound on function we want to sample from}

\item{\code{private$h_prims}}{A R6 namedVector that contains h'(x)}

\item{\code{private$h_vals}}{A R6 namedVector that contains h(x)}

\item{\code{private$s_cdfs}}{A R6 namedVector that contains each interval's CDF}

\item{\code{private$u_vals}}{A R6 namedVector that contains u(x)}

\item{\code{private$x}}{a vector of x values that are our sample values. Empty unless sample() run}

\item{\code{private$y}}{a vector of y values that begins to increase once sample() is run}

\item{\code{private$z}}{a vector of z values, or intersection points, that begins to increase once sample() is run}
}}

\section{Public Methods}{

\describe{
  \item{\code{f(x)}}{This method calculates our funx observation at a given x }
  \item{\code{plot_sampdist()}}{This method plots our envelope function }
  \item{\code{calc_sampdist()}}{This method calculates the sampling distribution from a uniform distribution}
  \item{\code{plot_samples()}}{This method plots our samples in a histogram}
  \item{\code{s()}}{This method calculates the envelope or sampling function }
  \item{\code{sample(n)}}{This method samples n points from using adaptive rejection sampling}}
}

\section{Private Methods}{

\describe{
  \item{\code{h(x)}}{This method calculates the log of our funcx at a given point x}
  \item{\code{init_hprim()}}{This method intializes our hprim values}
  \item{\code{init_hval()}}{This method initalizes our h values}
  \item{\code{init_scdf(}}{This method initalizes our cdf under the upper hull}
  \item{\code{init_u()}}{This method initializes our upper hull values}
  \item{\code{init_y()}}{This method initalizes our y values }
  \item{\code{l(p)}}{This method is a function that returns the slope of a lowerhull line given a point p}
  \item{\code{u(p)}}{This method is a function that returns the slope of a upperhull line given a point p}
  \item{\code{sampl_exph(}}{This method samples from our exp(h values)}
  \item{\code{update(y, hy, hy_prim)}}{This method updates our private variables if we reject a sample x_star}}
}

\examples{
ex <- ARS$new(dnorm, c(-Inf,Inf),mean=3, sd=1)
samples <- ex$sample(n=2000)
ex$plot_samples()
}
\keyword{datasets}
