% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.r
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling Function}
\usage{
ars(old_density_function, starting_points = NA, bounds = c(-Inf, Inf), n = 20)
}
\arguments{
\item{old_density_function}{Density to sample from}

\item{starting_points}{Starting points to begin adaptive sampling. Must be on either side of density maximum}

\item{bounds}{Bounds of density function. Must be provided when density is bounded e.g. gamma or beta}

\item{n}{Number of samples to draw}
}
\value{
n samples drawn from old_density_function
}
\description{
The overall function that computes all the steps in the Adaptive Rejection
Sampling. The function takes in a density function and has the option to
take in starting points, bounds and number of points to be sampled.
If the starting points and bounds are not specified the function sets
starting points and bounds. The default sampling size is 20 points.
The function returns an array with n sampled points from the given
density function.
}
\examples{
f <- function(x) dnorm(x) #set density function to sample from
samples <- ars(f, n=1000) 
hist(samples , probability = TRUE) #plot the samples
curve(f, add = TRUE, col = "blue") #plot the actual function 
}
