% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Generate random numbers using adaptive rejection sampling}
\usage{
ars(g, my_total_range, n)
}
\arguments{
\item{g}{the probability density distribution to sample from}

\item{my_total_range}{support of the probability density distribution}

\item{n}{number of observations.}
}
\description{
This function generate random numbers from given distribution using adaptive rejection sampling method.
}
\examples{
x = ars(dnorm, c(-Inf,Inf), n = 100)
}

