% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Generate random numbers using adaptive rejection sampling}
\usage{
ars(g, my_total_range, n)
}
\arguments{
\item{g}{the probability density distribution from which samples are drawn}

\item{my_total_range}{the user-defined domain of the probability density distribution}

\item{n}{desired number of observations to be returned}
}
\description{
This function generates random numbers from a user-defined distribution density function 
using the adaptive rejection sampling method.
}
\examples{
ars(dnorm, c(-Inf,Inf), n = 10000)
}

