% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_functions.R
\name{ars}
\alias{ars}
\title{Perform Adaptive Rejection Sampling}
\usage{
ars(n, x_abs = NULL, f, f_params = NULL, supp = c(-Inf, Inf))
}
\arguments{
\item{n}{The number of samples to return}

\item{x_abs}{A \code{numeric} vector of length \code{k} for \code{k > 1},
representing the abscissae.}

\item{f}{A function representing the target sampling distribution.}

\item{f_params}{A \code{list} of any associated parameters of \code{f}.}

\item{supp}{The support of \code{f}, as a two-membered \code{numeric} vector.
Default is \code{c(-Inf, Inf)}}
}
\value{
A named \code{list} containing \code{vals}: the collection of n
 samples, \code{k}: the count of abscissae added, \code{i}: the total number
 of samples attempted, and \code{m}: the amount of rejected samples.
}
\description{
\code{ars} calculates a given number of \code{n} random samples
from a given density \code{f}, initialized at the given abscissae
\code{x_abs}.
}
