% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(n, g = NA, D_left = -Inf, D_right = Inf, k = 10)
}
\arguments{
\item{n}{The number of desired samples}

\item{g}{The density function of interest such as `dnorm`
the valid parameters must be given, see examples.}

\item{D_left}{The desired left end of domain (optional), default = -Inf}

\item{D_right}{The desired right end of domain (optional), default = Inf}

\item{k}{The number of desired initial Abscissa (optional), default = 10}
}
\value{
A vector of sample from the targeted function provided by the user
}
\description{
An implementation of adaptive rejection sampling based on the tangent approach described
by Gilks and Wild (1992).
}
\examples{
dis_gamma <- function(x) {return(dgamma(x, shape = 2, rate = 2))}
ars(n=1000, g=dnorm , D_left = -3, D_right = 3, k = 20)
ars(n=1000, g=dis_gamma, D_left = -3, D_right = 3, k = 20)
}
