% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(n, g = NA, D_left = -Inf, D_right = Inf, k = 10)
}
\arguments{
\item{n}{The number of desired samples}

\item{g}{The density function of interest}

\item{D_left}{The desired left end of domain (optional), default = -Inf}

\item{D_right}{The desired right end of domain (optional), default = Inf}

\item{k}{The number of desired initial Abscissa (optional), default = 10}
}
\value{
Results from the adaptive rejection sample from the information provided by the user
}
\description{
An implementation of adaptive rejection sampling based on the tangent approach described
by Gilks and Wild (1992).
}
\examples{
ars(n=1000, g=dnorm , D_left = -3, D_right = 3, k = 20)

}
