% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars_function.R
\name{ARS}
\alias{ARS}
\title{The Adaptive Rejection Sampling Function}
\source{
Gilks, W. R. and Wild, P. (1992) Adaptive Rejection Sampling for Gibbs Sampling. Appl.Statist., 41 337 - 348.
}
\usage{
ARS(fx, n, sp = NA, l_bound = -Inf, u_bound = Inf)
}
\arguments{
\item{fx}{The log-concave function to be sampled from. Needs to be in the format of a function.}

\item{n}{Number of samples. Needs to be an integer.}

\item{sp}{Vector of Starting points. Needs to be at least length 2. Default to NA.}

\item{l_bound}{Value for lower bound. Needs to be a number. Defaults to -Inf.}

\item{l_bound}{Value for upper bound. Needs to be a number. Defaults to Inf.}
}
\value{
A vector of samples from f(x).
}
\description{
This function allows you to sample from any univariate log-concave probability density function f(x).
}
\examples{
ARS(dnorm, 100, l_bound = -10, u_bound = 10)

}
