% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive rejection sampler
The main ars() function}
\usage{
ars(h, n, domain, x1 = NULL, xk = NULL, x0 = 0, nmesh = 5,
  x_start = 0, max_x = 10000, min_step = 0.01, relax_factor = 10)
}
\arguments{
\item{h}{The log of the density function}

\item{n}{The number of points to be sampled}

\item{domain}{A vector of length 2 giving the domain (left bound, right bound) of the density function}

\item{x1}{The leftmost node of the abscissae, if not provided, the abscissae() will try to find one which has a suitable slope, leading to a unbiased and numerical-issues friendly sampling process}

\item{xk}{The the rightmost node of the abscissae, if not provided, the abscissae() will try to find one which has a suitable slope, leading to a unbiased and numerical-issues friendly sampling process}

\item{max_x}{See ?search}

\item{min_step}{See ?search}

\item{relax_factor}{See ?search}
}
\description{
Adaptive rejection sampler
The main ars() function
}
\examples{
h <- function (x){
log (dnorm (x, 0, 1))
}
samples <- ars(h, 10000, c(-Inf, Inf))
}

