% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARS.R
\name{ARS}
\alias{ARS}
\title{ARS}
\usage{
ARS(n, func, start = -Inf, end = Inf, fun_deriv = NULL, ...)
}
\arguments{
\item{n}{number of sample size.}

\item{start}{lower bound. (optional)}

\item{end}{upper bound. (optional)}

\item{fun_deriv}{the derivative of the density function. (optional)}

\item{...}{further arguments passed to or from other methods.}

\item{fun}{the density function of interest, not need to be normalized.}
}
\value{
a vector of sample from the density function with length n
}
\description{
Adaptive rejection sampling
}
\examples{

#Sample from normal distribution
n <- 100
ARS(n, dnorm)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

