% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{ars}
\usage{
ars(n, func, start = -Inf, end = Inf, hfun_deriv = NULL, ...)
}
\arguments{
\item{n}{number of sample size.}

\item{func}{the density function of interest.}

\item{start}{lower bound of the density function.}

\item{end}{upper bound of the density function.}

\item{hfun_deriv}{the derivatives of log density function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector of sample from the density function with length n
}
\description{
Adaptive rejection sampling
}
\examples{

#Sample from normal distribution
n <- 100
ARS(n, dnorm)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

