% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ARS}
\alias{ARS}
\title{ARS}
\usage{
ARS(num_samples, dens, bounds, deriv_func = NULL, abcissae = NULL)
}
\arguments{
\item{num_samples}{integer -> The sample size.}

\item{dens}{function -> The target density. Must be a continuous log-concave function
that takes one input as a float and returns the density as a float. The density does
not need to be normalized.}

\item{bounds}{vector -> Bounds of target density. Must be a two-element vector
of the lower and upper bounds of the target density describing where the density
is non-zero.}

\item{deriv_func}{(Optional) function -> Custom derivative function. Must be a
function that takes one input as a float and returns the derivative of the log
of the target density.}

\item{abcissae}{(Optional) vector -> Starting abcissae locations. Can give starting
abcissae locations. This can be helpful if you want the help the function find where
to start searching. If the lower bound is -Inf, you need at least one abcissae location
where the derivative of the logarithm of the density is positive. If the upper bound is
Inf, then you need at least one location where the derivative of the logarithm of the
density is negative. Requires at least 3 points.}
}
\value{
A vector of floats
}
\description{
This function generates samples from a custom distribution.
}
\examples{

norm_samp <- ARS(10000, dens = dnorm, bounds = c(-Inf,Inf))




set.seed(1)
f <- function(x) {
return(dnorm(x, 5, 3))
}
samples <- ARS(10000, dens = f, bounds = c(-Inf,Inf))
hist(samples, breaks = 20,freq = FALSE)
d <- dnorm(seq(-10,20,length.out=1000),5,3)
lines(seq(-10,20,length.out=1000),d)
ks.test(samples,rnorm(10000,5,3))

set.seed(1)
f <- function(x) {
    return(1000*dbeta(x,2,2))
}
samples <- ARS(10000, dens = f, bounds = c(0,1))

set.seed(1)
f <- function(x) {
    return(dunif(x, 5, 14))
}
samples <- ARS(10000, dens = f, bounds = c(5,14))

set.seed(1)
f <- function(x) {
    return(dexp(x, 2))
}
samples <- ARS(10000, dens = f, bounds = c(0,Inf))

set.seed(1)
f <- function(x) {
    return(dgamma(x, 10))
}
samples <- ARS(10000, dens = f, bounds = c(0,Inf))



set.seed(1)
f <- function(x) {
    return(dnorm(x, 5, 3))
}
d_func <- function(x){
    return(-(x-5)/3^2)
}
samples <- ARS(10000, dens = f, bounds = c(-Inf,Inf), deriv_func = d_func)




set.seed(1)
f <- function(x) {
return(dexp(x, 2))
}
samples <- ARS(10000, dens = f, bounds = c(0,Inf), abcissae = c(1,2,5))

}
\references{
Journal of the Royal Statistical Society. Series C (Applied Statistics), 1992, Vol.
41, No. 2 (1992), pp. 337-348
}
