% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(f = dnorm, nsamples = 100, min = -Inf, max = Inf,
  xinit = c(-1.5, -0.2, 0.5, 1))
}
\arguments{
\item{f}{A function of the continuous density we want to sample from. Defaults to standard normal.}

\item{nsamples}{The number of samples desired. Defaults to 100}

\item{min}{The minimum of the domain of \code{f}. Defaults to -Inf.}

\item{max}{The maximum of the domain of \code{f}. Defaults to Inf.}

\item{xinit}{Vector of starting points for which \code{f} is defined (at least 2).}
}
\value{
A vector (of length \code{nsamples}) of sampled points from the specified distribution.
}
\description{
Perform the adaptive rejection sampling from the specified log-concave density
}
\examples{
a <- ars(dnorm, nsamples = 100, min = -Inf, max = Inf, xinit = c(-1.5, -0.2, 0.5, 1))
g <- function(x) {dchisq(x, df = 3)}
}
