% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsFunction.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(g, n, lb = -Inf, ub = Inf, batchSize = 100, randomState = 1)
}
\arguments{
\item{n}{number of samples}

\item{lb}{lower bound on x axis}

\item{ub}{upper bound on x axis}

\item{batchSize}{number of seeds for inverse CDF, default = 100}

\item{randomState}{seed for set.seed()}

\item{G}{original function}
}
\value{
Samples from target density
}
\description{
This function is a statistical algorithm for generating samples from #' a univariate, log-concave density.
}
