% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Main function to carry out simulation}
\usage{
ars(n, g, D)
}
\arguments{
\item{n}{A number indicates how many accepted points user wants to generate.}

\item{g}{A function user wants to generate samples from.}

\item{D}{A vector with two elements indicates the domain of the sample 
generation.}
}
\value{
A numeric vector of length \code{n}, each element of which is
a value sampled from \eqn{g}.
}
\description{
Main function to carry out simulation
}

