% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(f, N, x0 = c(-1, 1), bounds = c(-Inf, Inf), ...)
}
\arguments{
\item{f}{vectorized function that's log-concave}

\item{N}{positive numeric; number of samples to generate}

\item{x0}{numeric vector; bounds of the sampling domain}

\item{bounds}{numeric vector of length 2; boundaries of the underlying sampling
distribution f(x)}

\item{...}{additional arguments to be passed to \code{f}}
}
\value{
A vector of N samples generated from the f(x) distribution.
}
\description{
Generates samples from a log-concave distribution via adaptive
rejection sampling.
}
\details{
If \code{x0} is not specified, it assumes the default value of \code{c(-1.0, 1.0)}. 
If \code{bounds} is not specified, it assumes the default value of \code{c(-Inf, Inf)}. 

The argument \code{f} must be a vectorized, log-concave function.

\code{x0} can be a numeric vector of any length, as long as the values are 
valid for the given \code{f}.
The \code{bounds} must be defined for \code{f}. For example, setting 
\code{bounds = c(-Inf, Inf)} will cause an error for exponential distributions 
because they are not defined on (-Inf, 0).

If the value of \code{f(x)} is too small for the range of x given by \code{bounds}, 
the program will throw an error.
}
\examples{
ars(dnorm, 100)
## generating 100 samples from Exp(0.5) distribution
ars(dexp, 100, x0 = 0.5, bounds = c(0, Inf), rate = 0.5)
}
\references{
Gilks and Wild.Adaptive Rejection Sampling for Gibbs Sampling(1992)

Paciorek.STAT243 Unit 9 - Simulation(2020)
}
