% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(f, N, x0 = c(-1, 1), bounds = c(-Inf, Inf), ...)
}
\arguments{
\item{f}{vectorized function that's log-concave}

\item{N}{positive numeric; number of samples to generate}

\item{x0}{numeric vector; bounds of the sampling domain}

\item{bounds}{numeric vector of length 2; boundaries of the underlying sampling
distribution f(x)}

\item{...}{further arguments to be passed to \code{f}}
}
\value{
A vector of N samples generated from the f(x) distribution.
}
\description{
Generates samples from a log-concave distribution via adaptive
rejection sampling.
}
