% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampler}
\usage{
ars(n, f, dfunc, D_min, D_max, verbose = FALSE)
}
\arguments{
\item{n}{the number of samples that you'd like taken from the distribution.}

\item{f}{the function used to determine the distribution from which you'll
sample}

\item{dfunc}{the derivative of the function.}

\item{D_min}{the minimum boundary limit of the distribution.}

\item{D_max}{the maximum boundary limit of the distribution.}

\item{verbose}{Whether you want R to print out the process as it goes through
ARS. This can be useful for longer computations.}
}
\description{
Sample points from a distribution using an adaptive rejection sampler.
}
\details{
Adaptive rejection sampling can be beneficial when sampling from
certain distributions by reducing the number of evaluations that must occur.
It reduces the evaluations by assuming log-concavity for any input function,
and because it doesn't need to update for new envelope and squeeze functions
every iteration. ARS is particularly useful in Gibbs Sampling, where
calculations are complicated but usually log-concave.
}
\examples{
# testing on normal distribution
h = function(y) {
return(-log(2 * pi * sigma ^ 2) / 2 - (y - mu) ^ 2 / (2 * sigma ^ 2))
}

# derivative of h
dh = function(y){
  return(- (y - mu)/sigma^2)
}

res_samples <- ars(n = 1000, f = h, dfunc = dh, D_min = -10, D_max = 30)


# Check variance and mean
mean(res_samples)
var(res_samples)

}
