% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampler}
\usage{
ars(n, f, dfunc, T_start, D_min, D_max, verbose = FALSE)
}
\arguments{
\item{n}{describe}

\item{f}{describe}

\item{dfunc}{describe}

\item{T_start}{describe}

\item{D_min}{describe}

\item{D_max}{describe}

\item{verbose}{describe

Details here

Examples here}
}
\description{
Description here
}
