% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(f, n, bounds = c(-Inf, Inf), x_init = NA)
}
\arguments{
\item{f}{function from which to generate samples from}

\item{n}{numbers of desired samples}

\item{bounds}{vector of length 2 defining the lower and upper bound of of the distribution f}

\item{x_init}{initial point within the bounds that used to initialize the abscissae points}
}
\value{
returns a vector of n samples from the distribution function f
}
\description{
This function implements an adaptive rejection sampling algorithm to generate n samples from any univariate log-concave function.
}
\details{
\code{ars}
}
\examples{
## sample 1000 points from a N(0, 1) distribution
ars(dnorm, n = 10, bounds = c(-10,10), x_init = 1)


## plot 5000 points from a Gamma(3, 4) distribution
gam_samps <- ars(f = function(x) {dgamma(x, 3, 4)}, n = 5000, bounds = c(0, Inf), x_init = 1)
hist(gam_samps)

}
\references{
Gilks, W. R., & Wild, P. (1992). Adaptive rejection sampling for Gibbs sampling.
\emph{Journal of the Royal Statistical Society: Series C (Applied Statistics)}, 41(2), 337-348.
}
\author{
Jeffrey Kuo, Hsiang-Chuan Sha, Tessa Weiss
}
