% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive-Rejection Sampling}
\usage{
ars(n, g = NA, h = NA, h_prime = NA, D_left = -Inf,
  D_right = Inf, k = 10, step = 3, center = 0)
}
\arguments{
\item{n}{The number of desired samples}

\item{g}{The density function of interest}

\item{h}{The log of g (users can provide either g or h)}

\item{h_prime}{H prime function (optional)}

\item{D_left}{The desired left end of domain (optional), default = -Inf}

\item{D_right}{The desired right end of domain (optional), default = Inf}

\item{k}{The number of desired initial abscissae (optional), default = 10}

\item{step}{The desired width (optional), default = 3}

\item{center}{the desired center (optional), default = 0}
}
\value{
Results from the adaptive rejection sample from the information provided by the user
}
\description{
The objective of this package is to create a function that
performs adaptive-rejection sampling based on the tangent approach described 
by Gilks et al. (1992).
}
\examples{
final_sample <- ars(n=1000, g=dnorm , D_left = -Inf, D_right = Inf, k = 20)
hist(final_sample, breaks = 50)

}
