% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampler}
\usage{
ars(g, a, b, N, n_per_step = 500)
}
\arguments{
\item{g}{input density function}

\item{a}{lower bound of function domain D (could be -Inf)}

\item{b}{upper bound of function domain D (could be Inf)}

\item{N}{number of observations required}

\item{n_per_step}{Default = 500. Number of observations between
each update of the hull function}
}
\value{
A one-dimensional vector of N independent observations from the the given input density
}
\description{
Sample from a log-concave density function，potentially unnormalized,
using the Adaptive Rejection Sampling method described in Gilks et al (1992)
}
\examples{
g <- function (x) {
return(dnorm(x))
}
ars_results <- ars(g, a = -Inf, b= Inf, N = 100000)

}
