% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(samp_n, g, a = -Inf, b = Inf)
}
\arguments{
\item{g}{Univariate log-concave probability density function f(x).}

\item{a}{First starting point where f(x) is defined, default value is -Inf.}

\item{b}{Second starting point where f(x) is defined, default value is Inf.}

\item{sample_n}{A number, denoting the number of sample points generated.}
}
\value{
A list containing Information about:
\item{f}{logf(x)}
\item{sample}{A vector of size N containing sample points}
}
\description{
Adaptive Rejection Sampling
}
\references{
Gilks, W.R., P. Wild. (1992) Adaptive Rejection Sampling for Gibbs Sampling, Applied Statistics 41:337–348.
}
\author{
Jue Wang, Li Yang, Winnie Gao
}
