% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling (ars) Function}
\usage{
ars(g, n, lb = -Inf, ub = Inf)
}
\arguments{
\item{g}{probability density function, as a function of x (e.g. g <- function(x) dnorm(x,0,1))}

\item{n}{number of values the final sample should contain}

\item{lb}{lower bound to evaluate the density on}

\item{ub}{upper bound to evaluate the density on}
}
\value{
sample of with specified (n) elements
}
\description{
Adaptive rejection sampling function that creates a sample
from a density function based on the Gilks(1992) paper.
}
