% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(target_density, n, x_domain= c(-Inf,Inf), num_iter_allowed=10*n, ...)
}
\arguments{
\item{target_density}{density function from which to sample (must be log-concave)}

\item{n}{number of samples}

\item{x_domain}{(Optional) domain of the density function from which to sample, default is the whole real line}

\item{num_iter_allowed}{(Optional) maximum number of iterations for the algorithm, default is 10*n}

\item{...}{additional arguments to pass to g such as mean and standard deviation}
}
\value{
returns (length n) vector of samples from the density function g(x)
}
\description{
\code{ars} returns a sample of length n from the density function g
}
\details{
Adaptive Rejection Sampling algorithm for sampling from any univariate log-concave density function g(x).
This algorithm is particularly useful in situations where computing g(x) may be difficult, where n
is the total number of samples.  Here, g(x) need not be normalized.
}
\examples{

'Standard normal distribution'
samps = ars(dnorm, 10000, mean=0, sd=1)

'Truncated normal distribution'
samps = ars(dnorm, 10000, c(-5, 5), mean=-3, sd=3)

'Beta distribution'
samps = ars(dbeta, n=3000, x_domain=c(0, 1), shape1=6, shape2=19)

'Logistic distribution'
samps = ars(dlogis, n=3000, x_domain=c(-Inf, Inf), location = 5, scale =2)

# Can also use hist to visualize distribution, e.g.
# hist(samps, breaks = 100, freq=FALSE)

}
\references{
Section 2.2 of Gilks et al. (1992)
}
\author{
Zhengpu Zhao, William Hymer, Dheeraj Prasanth
}
