% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adapative Rejection Sampling}
\usage{
ars(fn, min_bound, max_bound, p, n_iter = 1000)
}
\arguments{
\item{fn}{function; log-concave function f(x).}

\item{min_bound}{numeric; the lower bound of the domain as the first starting point. The default is -Inf.}

\item{max_bound}{numeric; the upper bound of the domain as the second starting point. The default is Inf.}

\item{p}{numeric vector; starting fixed points for evaluation.}

\item{n_iter}{numeric; number of points to sample. The default is 1000.}
}
\value{
\code{ars} returns a vector of length n_iter containing sampled values.
}
\description{
`ars()` generates random numbers from log-concave functions by adapative rejection sampling.
}
\details{
This function takes a numeric argument n and a function argument f
to generate random numbers from the given distribution in the specified sample size,
based on adapative rejection sampling method. For more detials, see Gilks and Wild (1992).
}
\examples{
##  Sample from Standard Normal Distribution
ars(dnorm, min_bound = -5, max_bound = 5, p = c(-1, 1))

## Sample from Gamma Distribution
gamma_test <- function(x){
k <- dgamma(x, shape=7.5, scale=1)
return(k)
}
ars(gamma_test, min_bound = 0.1, max_bound = 20, p = c(3,10), n_iter = 10000)

}
\references{
W. R. Gilks, P. Wild. (1992), "Adaptive Rejection Sampling for Gibbs Sampling," Applied Statistics 41:337–348.
}
\author{
Pan, Yanting; Ye, Zhenni; Myers, Vince; based on Gilks and Wild (1992).
}
