% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adaptive Rejection Sampling}
\usage{
ars(f, N, x0 = c(-1.0, 1.0), bounds = c(-Inf, Inf), ...)
}
\arguments{
\item{f}{function from which to sample (must be log-concave)}

\item{N}{number of observations desired}

\item{x0}{vector of initial x points used to approximate h=log(f(x))}

\item{bounds}{vector (length 2) of lower and upper bounds for the distribution function f}

\item{...}{Additional arguments to pass to f}
}
\value{
vector of samples from f(x).
}
\description{
\code{ars} returns a sample x of length N from the density function f(x)
}
\details{
Adaptive Rejection Sampling algorithm for sampling from any log-concave function f(x).
}
\examples{
N <- 10000
x <- ars(dnorm, N)
x <- ars(dexp, N, x0=c(1.0), bounds=c(0.0, Inf))
x <- ars(dgamma, N, x0=c(2.5), bounds=c(0.0, Inf), shape=3.0, rate=2.0)
## We could plot this to check
## hist(x, breaks = 50, freq=FALSE)

\dontrun{
ars(dgamma, N, x0=c(0.1), bounds=c(0.0, Inf), shape=3.0, rate=2.0)
}
}
\references{
Gilks, W. R., et al. (1995-01-01). "Adaptive Rejection Metropolis Sampling within Gibbs Sampling".
    \emph{Journal of the Royal Statistical Society. Series C (Applied Statistics).} \strong{44} (4): 455–472.
}
\seealso{
Wikipedia entry on \href{https://en.wikipedia.org/wiki/Rejection_sampling}{Rejection Sampling}.
}
\author{
Jonathan Morrell, Ziyang Zhou, Vincent Zijlmans
}
