% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_positive}
\alias{check_positive}
\title{check_positive}
\usage{
check_positive(f, var_min, var_max)
}
\arguments{
\item{f}{a function to check whether it is positive on the y axis}

\item{var_min}{the lower bound domain to check with f}

\item{var_max}{the upper bound domain to check with f}
}
\value{
a boolean TRUE or FALSE. A TRUE means the function is only positive
}
\description{
Check whether a given function is only positive
}
\examples{
check_positive(dnorm, -5, 5)
}
