% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh.R
\name{dh}
\alias{dh}
\title{Derivative of logarithm of functions}
\usage{
dh(x, g)
}
\arguments{
\item{x}{sorted points in T, which contains k abscissae in the domain D}

\item{g}{the density function used to perform the rejection sampling, where g(x) = cf(x). f(x) is the density that we actually want to sample points from}
}
\description{
This function calculate the derivative of h(x), where h(x) is the log of functions
}

