% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_l.R
\name{evall}
\alias{evall}
\title{eval_l}
\usage{
evall(x1, Tk, hfun.x)
}
\arguments{
\item{x1}{points that want to evalue lower bound.}

\item{Tk}{abscissae.}

\item{hfun.x}{a vector which includes all values of the log density function at those points in Tk.}
}
\value{
a vector. Lower bound value of the log density function for squeezing test.
}
\description{
Evaluate the lower bound value
}
\examples{

#standard normal distribution
hfun <- function(x) dnorm(x, log=TRUE)
h <- 1e-8
hfun_deriv <- function(x) (hfun(x + h) - hfun(x - h))/(2*h)

#Construct mat
Tk <- c(-1, 1)
hfun.x <- hfun(Tk)

#evaluate lower bound for -0.5
evall(-0.5, Tk, hfun.x)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

