% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalu.R
\name{evalu}
\alias{evalu}
\title{evalu}
\usage{
evalu(x.temp, u)
}
\arguments{
\item{x.temp}{points that want to evalue upper bound.}

\item{u}{the output from \code{genu} function. A list which first element is the intercept and slope of the tangent lines of log density function and second element is the intersection of these tangent lines.}
}
\value{
a vector. Upper bound value of the log density function for squeezing test.
}
\description{
Evaluate the upper bound value
}
\examples{

#standard normal distribution
hfun <- function(x) dnorm(x, log=TRUE)
h <- 1e-8
hfun_deriv <- function(x) (hfun(x + h) - hfun(x - h))/(2*h)

#Construct mat
Tk <- c(-1, 1)
mat <- cbind(Tk, hfun(Tk), hfun_deriv(Tk))
mat

#Generate intercept, slope and breaks
u <- genu(mat, start=-Inf, end=Inf)

#evaluate upper bound for -0.5
evalu(-0.5, u)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

