% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_CheckLogConcavity.R
\name{faux_CheckLogConcavity}
\alias{faux_CheckLogConcavity}
\title{Helper function to check the log concavity}
\usage{
faux_CheckLogConcavity(inp_gfun, inp_Dvec)
}
\arguments{
\item{inp_gfun}{A function user wants to generate samples from. This function is 
used to calculate \eqn{h(x)=\log(g(x))}{h(x) = ln(g(x))}.}

\item{inp_Dvec}{A numeric vector of length 2 indicating the support of
\code{inp_gfun}.}
}
\value{
A logical vector of length 1:
   \code{TRUE} if \code{inp_gfun} is log-concave,
   \code{FALSE} if \code{inp_gfun} is not log-concave.
}
\description{
Helper function to check the log concavity
}

