% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_CheckLogConcavity.R
\name{faux_CheckLogConcavity}
\alias{faux_CheckLogConcavity}
\title{Helper function to check the log concavity}
\usage{
faux_CheckLogConcavity(inp_xvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of 3 consecutive x values $(x_{1}, x_{2}, x_{3})$
in the domain of the user supplied function $g(x)$}

\item{inp_gfun}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x)). We can then test the log-concavity
at the 3 x-coordinates $(x_{1}, x_{2}, x_{3})$}
}
\value{
boolean value TRUE or FALSE to indicate the log concavity
}
\description{
Helper function to check the log concavity
}

