% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_findmode.R
\name{faux_findmode}
\alias{faux_findmode}
\title{Helper function to find mode of a given univariate function \eqn{g(x)}}
\usage{
faux_findmode(optim_intervalvec, inp_gfun)
}
\arguments{
\item{optim_intervalvec}{A vector with two elements indicates the support domain of
the sample generation.}

\item{inp_gfun}{A function of x which the user wants to generate samples from.
This function is used to calculate \eqn{h(x)=\log(g(x))}.}
}
\value{
two starting points.
}
\description{
Helper function to find mode of a given univariate function \eqn{g(x)}
}

