% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_hx.R
\name{faux_hx}
\alias{faux_hx}
\title{Helper function to get h(x) = ln(g(x)).}
\usage{
faux_hx(inp_gfun)
}
\arguments{
\item{inp_gfun}{A function of x which the user wants to generate samples from.
This function is used to calculate h(x)=ln(g(x)).}
}
\value{
the h(x)
}
\description{
Helper function to get h(x) = ln(g(x)).
}

