% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_InitChoose.R
\name{faux_InitChoose}
\alias{faux_InitChoose}
\title{Helper function to choose two starting points}
\usage{
faux_InitChoose(inp_gfun, inp_Dvec, inp_Initnumsampvec = 2)
}
\arguments{
\item{inp_gfun}{A function user wants to generate samples from. This function
is used to calculate \eqn{h(x)=\log(g(x))}}

\item{inp_Dvec}{A vector with two elements indicates the support domain of
the sample generation.}

\item{inp_Initnumsampvec}{An even integer determining the number of points to
initially sample - should be even}
}
\value{
two starting points.
}
\description{
Helper function to choose two starting points
}

