% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Lkx.R
\name{faux_Lkx}
\alias{faux_Lkx}
\title{Helper function to get the lower bound linear function $l_{k}(x)$}
\usage{
faux_Lkx(inp_xvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of x values of all points and we should be able to get the
index of x}

\item{inp_gfun}{A function user wants to generate samples from. This function is
use to calculate h(x)=ln(g(x))}
}
\value{
the lower bound linear function $l_{k}(x)$ for x in the interval
[$x_{j}$,$x_{j+1}$]
}
\description{
Helper function to get the lower bound linear function $l_{k}(x)$
}

