% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_SampleSkx.R
\name{faux_SampleSkx}
\alias{faux_SampleSkx}
\title{Helper function to sample a value x*from $s_{k}(x)$}
\usage{
faux_SampleSkx(inp_xvec, inp_zvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of x values of all points}

\item{inp_zvec}{A vector of z values of all points and we should be able to get
the index of z}

\item{inp_gfun}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}
}
\value{
a x value sampling from function $s_{k}(x)$
}
\description{
Helper function to sample a value x*from $s_{k}(x)$
}

