% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_SampleSkx.R
\name{faux_SampleSkx}
\alias{faux_SampleSkx}
\title{Helper function to sample a value \eqn{x^*} from \eqn{s_{k}(x)}}
\usage{
faux_SampleSkx(inp_uintervallist, inp_sfunlist)
}
\arguments{
\item{inp_uintervallist}{A list of intervals between the \eqn{z} points}

\item{inp_sfunlist}{A list of functions which form the \eqn{s_{k}(x)}
function}
}
\value{
a value sampled from the function \eqn{s_{k}(x)}
}
\description{
Helper function to sample a value \eqn{x^*} from \eqn{s_{k}(x)}
}

