% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Skx.R
\name{faux_Skx}
\alias{faux_Skx}
\title{Helper function to create piecewise function \eqn{s_{k}(x)}{sk(x)}}
\usage{
faux_Skx(inp_uintervallist, inp_ufunlist)
}
\arguments{
\item{inp_uintervallist}{A list of intervals between \eqn{z} values, as in the 
output from uInterval.}

\item{inp_ufunlist}{A list of functions, the output from uFun.}
}
\value{
A list of functions.
The length of the list is equal to the length of the inputs
\code{inp_uintervallist} and \code{inp_ufunlist}.
Each element of the list is one piece of the piecewise function
\eqn{s_{k}(x)}{sk(x)}.
}
\description{
Helper function to create piecewise function \eqn{s_{k}(x)}{sk(x)}
}

