% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Skx.R
\name{faux_Skx}
\alias{faux_Skx}
\title{Helper function to create piecewise function {s_k(x)}}
\usage{
faux_Skx(inp_uintervallist, inp_ufunlist)
}
\arguments{
\item{inp_uintervallist}{A list of intervals between z values, as in the output
from uInterval.}

\item{inp_ufunlist}{A list of functions, the output from uFun.}
}
\value{
the function $s_{k}(x)$ for all x in the domain of g, in list format,
where each element of the list is one element of the piecewise function.
}
\description{
Helper function to create piecewise function {s_k(x)}
}

