% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_uInterval.R
\name{faux_uInterval}
\alias{faux_uInterval}
\title{Helper function to create the z intervals.}
\usage{
faux_uInterval(inp_z)
}
\arguments{
\item{inp_z}{A vector of z values, such as the output from faux_Zj().}
}
\value{
A list, with each element of the list an interval between 2 z points
}
\description{
Helper function to create the z intervals.
}

