% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_uInterval.R
\name{faux_uInterval}
\alias{faux_uInterval}
\title{Helper function to create the \eqn{z} intervals.}
\usage{
faux_uInterval(inp_z)
}
\arguments{
\item{inp_z}{A vector of \eqn{z} values, such as the output from
\code{faux_Zj()}.}
}
\value{
A list of numeric vectors, each of length 2.
The length of the list is one less than the length of the input \code{inp_z}.
Each element of the list is an interval between 2 consecutive \eqn{z} points.
}
\description{
Helper function to create the \eqn{z} intervals.
}

