% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_Ukx.R
\name{faux_Ukx}
\alias{faux_Ukx}
\title{Helper function to get the upper bound linear function $u_{k}(x)$}
\usage{
faux_Ukx(inp_xvec, inp_zvec, inp_gfun)
}
\arguments{
\item{inp_xvec}{A vector of x values of all points}

\item{inp_zvec}{A vector of z values of all points and we should be able to get 
the index of z}

\item{inp_gfun}{A function user wants to generate samples from. This function is 
used to calculate h(x)=ln(g(x))}
}
\value{
the upper bound linear function $u_{k}(x)$ for x in the interval 
[$z_{j-1}$,$z_{j}$]
}
\description{
Helper function to get the upper bound linear function $u_{k}(x)$
}

