% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_Zj.R
\name{faux_Zj}
\alias{faux_Zj}
\title{Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$}
\usage{
faux_Zj(inp_x, inp_gfun)
}
\arguments{
\item{inp_gfun}{A function user wants to generate samples from. This function is
used to calculate h(x)=ln(g(x))}

\item{inp_xvec}{A vector of x values of your points. The first element in the vector
has smaller x-coordinate.}
}
\value{
intersection of tangents of $x_{j}$ and $x_{j+1}$ named $z_{j}$
}
\description{
Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$
}

