% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faux_Zj.R
\name{faux_Zj}
\alias{faux_Zj}
\title{Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$}
\usage{
faux_Zj(inp_xvec, inp_gfun, inp_Dvec)
}
\arguments{
\item{inp_xvec}{A vector of x values of your points. The vector is ordered in
an increasing order.}

\item{inp_gfun}{A fucntion user wants to generate samples from. This function
is used to calculate h(x)=ln(g(x))}

\item{inp_Dvec}{A vector with 2 elements indicating the domain the function g}
}
\value{
a vector of intersection of tangents
}
\description{
Helper function to get the intersection of tangents at $x_{j}$ and $x_{j+1}$
}

