% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zj.R
\name{get_zj}
\alias{get_zj}
\title{Intersection points zj}
\usage{
get_zj(xj, hj, slopes)
}
\arguments{
\item{xj}{sorted points in T, which contains k abscissae in the domain D}

\item{hj}{evaluated h at xj, where h = log g(x). g is the density function used to perform the rejection sampling}

\item{slopes}{slope of h evaluated at xj; they are the outputs of dh function calculated at x1 through xk}
}
\description{
This function returns the intersection points generated from piecewise linear upper hull
}
\examples{
get_zj(xj, hj, slopes)
get_zj(c(1,2,3), c(4,8,12), c(1, 0,-1))
}

