% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitTk.R
\name{initTk}
\alias{initTk}
\title{InitTk}
\usage{
initTk(hfun, hfun_deriv, start, end)
}
\arguments{
\item{hfun}{log of the density function.}

\item{hfun_deriv}{the derivatives of log density function.}

\item{start}{lower bound of the density function.}

\item{end}{upper bound of the density function.}
}
\value{
a k by 3 matrix which its first column are initial points, second column are log of the density function and third column are the derivatives of log density function. k is the total number of initial points.
}
\description{
Generate the initial value for Adaptive rejection sampling
}
\examples{

#Initial value for standard normal distribution
set.seed(0)
hfun <- function(x) dnorm(x, log=TRUE)
h <- 1e-8
hfun_deriv <- function(x) (hfun(x + h) - hfun(x - h))/(2*h)
InitTk(hfun, hfun_deriv, -Inf, Inf)
}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

