% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lowerHull}
\alias{lowerHull}
\title{A function that, given a vector x, returns the corresponding values of the upper hull function}
\usage{
lowerHull(x, abscissae_summary)
}
\arguments{
\item{x}{a numeric vector consists of points on the x-axis}

\item{abscissae_summary}{a matrix with the three columns being x, h(x), and h'(x) at the support points respectively}
}
\value{
U: a numeric vector that contains the corresponding upper hull values, with same length as x
}
\description{
A function that, given a vector x, returns the corresponding values of the upper hull function
}
\examples{
h <- function(x){
  return(log(dnorm(x)))
}

xSupport <- seq(-3, 3, length.out = 10)

abscissae_summary = abscissaeSummary(xSupport,h)

x <- seq(-3, 3, length.out = 10)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)

x <- seq(-3, 3, length.out = 30)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)
}

