% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_x.R
\name{samplex}
\alias{samplex}
\title{sample_x}
\usage{
samplex(u, y)
}
\arguments{
\item{u}{list contains coefficients and breaks}

\item{y}{uniform random number}
}
\value{
x value
}
\description{
Sample the x value from S(x)
}
\examples{

}
\author{
Chao Mao, Xian Shi, Chih-Hui Wang, Luyun Zhao
}

