% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk.R
\name{sk}
\alias{sk}
\title{Generate a time series of fractional Brownian motion.}
\usage{
sk(x, xi, yi, k, my_domains)
}
\arguments{
\item{hurst}{the hurst index, with the default value 0.71}

\item{n}{the number of points between 0 and 1 that will be generated, with the default value 100}
}
\description{
This function generatea a time series of one dimension fractional Brownian motion.
adapted from http://www.mathworks.com.au/matlabcentral/fileexchange/38935-fractional-brownian-motion-generator .
}
\examples{
fbm()
plot(fbm())
d <- fbm(hurst=0.2, n=1000)
plot(d)
}

