% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_piecewise.R
\name{upper_piecewise}
\alias{upper_piecewise}
\title{Piecewise linear upper hull}
\usage{
upper_piecewise(x, xj, hj, slopes, domains)
}
\arguments{
\item{xj}{sorted points in T, which contains k abscissae in the domain D}

\item{hj}{evaluated h at xj, where h = log g(x). g is the density function used to perform the rejection sampling}

\item{slopes}{slope of h evaluated at xj; they are the outputs of dh function calculated at x1 through xk}

\item{domains}{domain D, which contains the lower bound and the upper bound}
}
\description{
This function returns the piecewise linear upper hulls from the tangents to h(x) at the abscissae xj in Tk
}
\examples{
upper_piecewise(x, x_intercepts, hj, slopes, leftb, rightb)
upper_piecewise(4,c(1,23,3),c(2,1,4),c(-1,0,10),c(-5,5))
}

