% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{upperHull}
\alias{upperHull}
\title{A function that, given a vector x, returns the corresponding function values from the upper hull}
\usage{
upperHull(x, abscissae_summary)
}
\arguments{
\item{x}{a numeric vector consists of points on the x-axis}

\item{abscissae_summary}{a matrix with the three columns being x, h(x), and h'(x) at the support points respectively}
}
\value{
U: a numeric vector that contains the corresponding upper hull values, with same length as x
}
\description{
A function that, given a vector x, returns the corresponding function values from the upper hull
}
\examples{
# Test1: log of normal distribution
h <- function(x){
  return(log(dnorm(x)))
}

xSupport <- seq(-3, 3, length.out = 10)

abscissae_summary = abscissaeSummary(xSupport,h)

x <- seq(-3, 3, length.out = 10)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)

x <- seq(-3, 3, length.out = 30)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)

h <- function(x){
 return(log(dchisq(x,df = 2)))
}

xSupport <- seq(-3, 3, length.out = 10)
abscissae_summary = abscissaeSummary(xSupport,h)
x <- seq(-3, 3, length.out = 10)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)

# Test2: log of uniform distribution
h <- function(x){
 return(log(dunif(x,0,1)))
}

xSupport <- seq(-3, 3, length.out = 100)
abscissae_summary = abscissaeSummary(xSupport,h)
x <- seq(-3, 3, length.out = 10)
x <- seq(0.2, 0.9, length.out = 10)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)

# Test3: log of piecewise constant distribution
h <- function(x){
 a= sapply(x,function(x){if(x<1) return (x)
 else if(x>=1 & x <=2) return (1)
 else if(x>2) return (-x+3)})
 return (a)
}

xSupport <- seq(-3, 3, length.out = 15)
abscissae_summary = abscissaeSummary(xSupport,h)
x <- seq(-3, 3, length.out = 10)
x <- seq(0.2, 0.9, length.out = 10)
upperHull(x,abscissae_summary)
lowerHull(x,abscissae_summary)
}

